Print all the elements of a list_t list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 size_t print_list(const list_t *h)
 {
 	const list_t *tmp = h;
-	int num = 0;
+	size_t num = 0;
 
 	while (tmp != NULL)
 	{
